fix: change some logics

+++ processor/processor_clipreid_stage2.py
@@ -89,15 +89,13 @@ def do_train_stage2(cfg,
             target = vid.to(device)
             if cfg.MODEL.SIE_CAMERA:
                 target_cam = target_cam.to(device)
-            else: 
+            else:
                 target_cam = None
             if cfg.MODEL.SIE_VIEW:
                 target_view = target_view.to(device)
-            else: 
+            else:
                 target_view = None
-            # with amp.autocast(enabled=True):
-            #     text_features = model(label = target, get_text = True)
-            with amp.autocast(enabled=True, dtype=torch.float16):
+            with amp.autocast(enabled=True):
                 score, feat, image_features = model(x = img, label = target, cam_label=target_cam, view_label=target_view)
                 logits = image_features @ text_features.t()
                 loss = loss_fn(score, feat, target, target_cam, logits)
